requirements.txt
Package             Version
------------------- -------
asgiref             3.8.1
Django              5.0.7
djangorestframework 3.15.2
pip                 24.1.2
sqlparse            0.5.0
tzdata              2024.1
